fix: use server action to encrypt & decrypt
