+++ .env.example
@@ -2,4 +2,4 @@ SECRET_KEY=12345678
 BASE_URL=http://localhost:3000
 SUPABASE_URL=http://127.0.0.1:54321
 SUPABASE_KEY=
-NEXT_PUBLIC_CYPHER_KEY=b28cf95d85b0a9c825b0010c4f902adf # Generate a secret key with 32 bytes length (256 bits)
+CYPHER_KEY=b28cf95d85b0a9c825b0010c4f902adf # Generate a secret key with 32 bytes length (256 bits)
